+++ src/main/java/com/sparta/kidscafe/api/kakao/MapService.java
@@ -1,4 +1,4 @@
-package com.sparta.kidscafe.api.map;
+package com.sparta.kidscafe.api.kakao;
 
 import com.sparta.kidscafe.common.util.GeoUtil;
 import org.json.JSONArray;
@@ -22,10 +22,16 @@ public class MapService {
   private String restApiKey;
   private final RestTemplate restTemplate;
   private final GeoUtil geoUtil;
+  private final Regions regions;
 
   public MapService(RestTemplateBuilder builder, GeoUtil geoUtil) {
     this.restTemplate = builder.build();
     this.geoUtil = geoUtil;
+    regions = new Regions();
+  }
+
+  public String[][] getRegions() {
+    return regions.getRegions();
   }
 
   public Point convertAddressToGeo(String address) {
